Review: extract shared token-usage parsing into TokenUsageParsing module
Address code review feedback:
- Remove unnecessary `require "rubygems"` from github_copilot.rb
- Extract 10+ duplicated token-parsing methods (normalized_model_name,
  effective_model_name, nested_hash_value, normalize_token_count,
  token_count_for, select_best_usage_payload, usage_token_field_count,
  usage_token_total, usage_with_token_counts?, negative_token_count_present?)
  into Providers::TokenUsageParsing shared module
- Fix longest_common_substring_length threshold from >= 2 to a proportional
  threshold (min_length / 2, floor 1) so it provides a meaningful similarity
  signal
- Add depth guard (MAX_METRICS_DEPTH = 5) to model_metrics_usages recursion
