+++ lib/agent_harness.rb
@@ -252,6 +252,7 @@ require_relative "agent_harness/provider_health_check"
 require_relative "agent_harness/providers/registry"
 require_relative "agent_harness/providers/adapter"
 require_relative "agent_harness/providers/base"
+require_relative "agent_harness/providers/token_usage_parsing"
 require_relative "agent_harness/providers/anthropic"
 require_relative "agent_harness/providers/aider"
 require_relative "agent_harness/providers/codex"
